arrange files under core folder, extract storage into storage.py
